Updated the getURLForNode function
we first check if the node type is 'URL'. If it is, we directly return the content. If not, we try to get the child node with the name 'URL'. If there's no child or the child's type is not 'URL', we return undefined, indicating that the URL couldn't be found. This should prevent the function from returning null or causing errors when the URL is not found.